feat: add support for heregex templates
Fixes #557
Fixes #341

Most of the existing code in StringPatcher was also relevant for heregexes, so I
pulled the shared code into an InterpolatedPatcher class.

One oddity here is the way backslashes are handled in heregexes. You can use a
backslash to escape a whitespace character (i.e. to include it), but other than
that, backslashes just represent the backslash character. So when turning a
heregex into a string, we need to escape backslashes when not followed by
whitespace, but skip them if they are. To accomplish this, I generalized the
`escape` function to parameterize how it skips characters.

With this change, decaffeinate can convert all 1213 files in my codebase!
