+++ package.json
@@ -35,9 +35,9 @@
     "ast-processor-babylon-config": "^1.0.0",
     "automatic-semicolon-insertion": "^1.0.2",
     "babylon": "^6.12.0",
-    "coffee-lex": "^4.2.2",
+    "coffee-lex": "^4.2.3",
     "decaffeinate-coffeescript": "^1.10.0-patch5",
-    "decaffeinate-parser": "^3.0.0",
+    "decaffeinate-parser": "^4.0.0",
     "detect-indent": "^4.0.0",
     "esnext": "^3.0.6",
     "lines-and-columns": "^1.1.5",
