+++ test/regular_expression_test.js
@@ -13,14 +13,17 @@ describe('regular expressions', () => {
     check(`a.a(/#/)`, `a.a(/#/);`);
   });
 
-  it('rewrites block regular expressions as normal regular expressions', () => {
+  it('rewrites block regular expressions as a RegExp call with a multiline string', () => {
     check(`
       a = ///
         foo .*
         bar
       ///
     `, `
-      let a = /foo.*bar/;
+      let a = new RegExp(\`\\
+      foo.*\\
+      bar\\
+      \`);
     `);
   });
 
@@ -36,7 +39,29 @@ describe('regular expressions', () => {
     check(`/\\/\\//`, `/\\/\\//;`);
   });
 
-  it('escapes slashes in heregexes', () => {
-    check(`///a/b///`, `/a\\/b/;`);
+  it('escapes necessary characters in heregexes', () => {
+    check(`///a/b\`c\${d\\e///`, `new RegExp(\`a/b\\\`c\\\${d\\\\e\`);`);
+  });
+
+  it('properly handles flags in heregexes', () => {
+    check(`///a\nb///gi`, `new RegExp(\`a\\\nb\`, 'gi');`);
+  });
+
+  it('handles interpolations within heregexes', () => {
+    check(`
+      ///
+        foo  # hello #{abc}d
+        #{bar}
+      ///g
+    `, `
+      new RegExp(\`\\
+      foo\${abc}d\\
+      \${bar}\\
+      \`, 'g');
+    `);
+  });
+
+  it('allows escaping spaces in heregexes', () => {
+    check(`///a\\ b\\\tc\\\nd///`, `new RegExp(\`a b\tc\nd\`);`);
   });
 });
